+++ cli-tools/cli/bin/aofl.js
@@ -8,10 +8,10 @@ program
 .version(chalk.green(version))
 .on('--help', () => {
   console.log(`
-  Examples:
+Examples:
 
-    aofl --help
-    aofl source
+aofl --help
+aofl source
 `);
 })
 .command('init <path>', 'Initializes an aofl project')
@@ -19,7 +19,10 @@ program
 .command('conclude <moduleName>', 'conclude module')
 .command('g [type] [name]', 'scoffold component')
 .command('i18n <path>', 'Generates translation.json based on provided pattern and scoped by the presence of an i18n directory anywhere in the project.')
-.command('dom-scope [path]', 'validate generated dom-scope');
+.command('dom-scope [path]', 'validate generated dom-scope')
+.command('build', 'Builds your aofl project.')
+.command('test', 'Runs unit tests')
+.command('serve', 'Builds and servers your aofl project.');
 
 program
 .parse(process.argv);