+++ cli-tools/cli/lib/environment-enumerate.js
@@ -1,0 +1,5 @@
+module.exports = {
+  DEVELOPMENT: 'development',
+  PRODUCTION: 'production',
+  TEST: 'test'
+};