+++ cli-tools/cli/lib/webpack-config/__config/sw.js
@@ -1,0 +1,10 @@
+workbox.setConfig({
+  debug: false
+});
+
+workbox.skipWaiting();
+workbox.clientsClaim();
+
+self.__precacheManifest = [].concat(self.__precacheManifest || []);
+workbox.precaching.suppressWarnings();
+workbox.precaching.precacheAndRoute(self.__precacheManifest, {});